+++ src/font-chef/render-result.c
@@ -29,7 +29,7 @@ float fc_text_segment_width(struct fc_text_segment const * segment, struct fc_ch
   return mapping[segment->last].target.right - mapping[segment->first].target.left;
 }
 
-size_t fc_wrap(struct fc_character_mapping mapping[], size_t glyph_count, float line_width, float line_height, float space_width, enum fc_alignment aligment) {
+uint32_t fc_wrap(struct fc_character_mapping mapping[], size_t glyph_count, float line_width, float line_height, float space_width, enum fc_alignment aligment) {
   struct fc_text_segment * words = calloc(sizeof(*words), glyph_count);
   struct fc_text_segment * lines = calloc(sizeof(*lines), glyph_count);
   size_t word_count = 0;
